presets(glsl): split paper shader presets
